Refactoring the Core package
- now a single facade for core, handling both vectorized and standard
  environments
- Core dispatches appropriate subclass
- minor performance issue storing n_active_envs to avoid CPU/GPU sync
