+++ mushroom_rl/core/_impl/vectorized_core_logic.py
@@ -7,6 +7,7 @@ class VectorizedCoreLogic(CoreLogic):
         self._array_backend = ArrayBackend.get_array_backend(backend)
         self._n_envs = n_envs
         self._running_envs = self._array_backend.zeros(n_envs, dtype=bool)
+        self._n_active_envs = 0
 
         super().__init__()
 
@@ -41,6 +42,7 @@ class VectorizedCoreLogic(CoreLogic):
             mask[last] = new_mask
 
         self._running_envs = self._array_backend.copy(mask)
+        self._n_active_envs = mask.sum().item()
 
         return mask
 
@@ -53,10 +55,9 @@ class VectorizedCoreLogic(CoreLogic):
         return initial_state
 
     def after_step(self, last):
-        n_active_envs = self._running_envs.sum().item()
-        self._total_steps_counter += n_active_envs
-        self._current_steps_counter += n_active_envs
-        self._steps_progress_bar.update(n_active_envs)
+        self._total_steps_counter += self._n_active_envs
+        self._current_steps_counter += self._n_active_envs
+        self._steps_progress_bar.update(self._n_active_envs)
 
         completed = last.sum().item()
         self._total_episodes_counter += completed
@@ -67,6 +68,7 @@ class VectorizedCoreLogic(CoreLogic):
         super().after_fit(n_carry_forward_steps)
         if self._n_episodes_per_fit is not None:
             self._running_envs = self._array_backend.zeros(self._n_envs, dtype=bool)
+            self._n_active_envs = 0
             return self._array_backend.ones(self._n_envs, dtype=bool)
         else:
             return last
@@ -74,6 +76,7 @@ class VectorizedCoreLogic(CoreLogic):
     def _reset_counters(self):
         super()._reset_counters()
         self._running_envs = self._array_backend.zeros(self._n_envs, dtype=bool)
+        self._n_active_envs = 0
 
     @property
     def converter(self):
